+++ settings.py
@@ -76,7 +76,7 @@ class Plot(BaseSettings):
 
 
 class Settings(BaseSettings):
-    scale_factor: float = 1.0
+    scale_factor: int = 1
 
     paths: Paths = Paths()
     plot: Plot = Plot()
